[Ref] Silence Clang 10 warning about inexact result in implicit integral-to-floatingpoint conversions when the source value is std::numeric_limits::max().
git-svn-id: https://source.openmpt.org/svn/openmpt/branches/OpenMPT-1.27@13454 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ common/misc_util.h
@@ -810,7 +810,7 @@ namespace Util
 	{
 		static_assert(std::numeric_limits<T>::is_integer == true, "Type is a not an integer");
 		const double valRounded = Round(val);
-		MPT_ASSERT(valRounded >= (std::numeric_limits<T>::min)() && valRounded <= (std::numeric_limits<T>::max)());
+		MPT_ASSERT(valRounded >= static_cast<double>((std::numeric_limits<T>::min)()) && valRounded <= static_cast<double>((std::numeric_limits<T>::max)()));
 		const T intval = static_cast<T>(valRounded);
 		return intval;
 	}
@@ -819,7 +819,7 @@ namespace Util
 	{
 		static_assert(std::numeric_limits<T>::is_integer == true, "Type is a not an integer");
 		const float valRounded = Round(val);
-		MPT_ASSERT(valRounded >= (std::numeric_limits<T>::min)() && valRounded <= (std::numeric_limits<T>::max)());
+		MPT_ASSERT(valRounded >= static_cast<float>((std::numeric_limits<T>::min)()) && valRounded <= static_cast<float>((std::numeric_limits<T>::max)()));
 		const T intval = static_cast<T>(valRounded);
 		return intval;
 	}
